Fixed warnigns

+++ RSMToolkit/NSArray+RSMToolKit.m
@@ -11,7 +11,7 @@
 @implementation NSArray (RSMToolKit)
 
 - (NSArray*) deepCopy {
-    unsigned int count = [self count];
+    NSUInteger count = [self count];
     id cArray[count];
 	
     for (unsigned int i = 0; i < count; ++i) {
@@ -30,7 +30,7 @@
     return ret;
 }
 - (NSMutableArray*) mutableDeepCopy {
-    unsigned int count = [self count];
+    NSUInteger count = [self count];
     id cArray[count];
 	
     for (unsigned int i = 0; i < count; ++i) {
